cpu/stm32_common: unified UART driver
